requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf_spectacular

